Sample Game

+++ include/GameEngine/Core/Time.hpp
@@ -6,11 +6,11 @@
 namespace GameEngine {
 	class GEAPI Time {
 	public:
-		static double RunningTime();
-		static double DeltaTime();
-        static double FrameRate();
+		static float RunningTime();
+		static float DeltaTime();
+        static float FrameRate();
 		
-		static void Sleep(double millis);
+		static void Sleep(float millis);
 		static void BusySleep(uint32_t micros);
 	private:
 		Time() = delete;
